requirements.txt
# Python dependencies for AI-102 labs and code samples

# Azure AI Services SDKs
azure-ai-vision-imageanalysis>=1.0.0
azure-ai-textanalytics>=5.3.0
azure-cognitiveservices-speech>=1.37.0
azure-ai-translation-text>=1.0.0
azure-ai-formrecognizer>=3.3.0
azure-search-documents>=11.4.0
azure-ai-contentsafety>=1.0.0

# Azure OpenAI
openai>=1.30.0

# Azure Identity & Management
azure-identity>=1.16.0
azure-mgmt-cognitiveservices>=13.5.0
azure-keyvault-secrets>=4.8.0

# Azure Storage (for lab data)
azure-storage-blob>=12.19.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.26.0

# Development
httpie>=3.2.0
